Initial commit
Added the .NET Core MVC server for serving up the single page app
and the REST API for the SPA and Native apps to consume.

The build and run process is automated through the NPM scripts fo
both the client javascript and the server. Would probably be good to
separate those out if the server app takes too long to build or dev is done.